+++ .gitignore
@@ -147,3 +147,4 @@ tfrecords.003
 /train_images/
 /test_images/
 /LIGO_data/
+/TFRecords/
